+++ packages/radix-ui-themes/changelog.md
@@ -28,6 +28,10 @@
   - Update high contrast variant design
 - `Theme`:
   - [**Breaking**] Remove `textColor` prop
+  - [**Breaking**] Remove `backgroundColor` prop
+  - [**Breaking**] Rename `applyBackground` prop to `background`
+  - [**Breaking**] Should always be within `body` now (rather than wrapping `html` or `body`)
+  - [**Breaking**] When `appearance` is `dark` or `light`, we inject a script (similar to `next-themes`) to set the theme class / color-scheme on `html`. Because of this, `suppressHydrationWarning` needs to be added to `html`.
 - `ThemePanel`:
   - New design